% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics-handling.r
\name{add_eqv_aes}
\alias{add_eqv_aes}
\title{Add equivalent aesthetics to vector}
\usage{
add_eqv_aes(aes_names)
}
\arguments{
\item{aes_names}{a character vector of aesthetic names}
}
\value{
a character vector of aesthetic names including any localized or base
R equivalent argument names accepted by
\href{[https://github.com/tidyverse/ggplot2}{ggplot2}.
}
\description{
Many \href{[https://github.com/tidyverse/ggplot2}{ggplot2}
\code{\link[ggplot2]{Geom}} objects will accept a number of alternative
spellings and base R equivalent arguments as aesthetic parameters. This
function adds those equivalent values to a character vector of aesthetic
names.
}
\examples{
ggpackets:::add_eqv_aes(c('color', 'fill'))

}
\seealso{
\code{\link[=allowed_aesthetics]{allowed_aesthetics()}}, \code{\link[=filter_aesthetics]{filter_aesthetics()}},
\code{\link[=remove_aesthetics]{remove_aesthetics()}}, \code{\link[=flatten_aesthetics_to_group]{flatten_aesthetics_to_group()}}, \code{\link[=add_eqv_aes]{add_eqv_aes()}},
\code{\link[=split_aes_from_dots]{split_aes_from_dots()}}, \code{\link[=is_uneval]{is_uneval()}}
}
\keyword{aes}
\keyword{aesthetics}
