% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics-handling.r
\name{allowed_aesthetics}
\alias{allowed_aesthetics}
\title{Get Geom allowed aesthetics}
\usage{
allowed_aesthetics(geom = NULL)
}
\arguments{
\item{geom}{a \href{[https://github.com/tidyverse/ggplot2}{ggplot2}
\code{\link[ggplot2]{Geom}} object or NULL}
}
\value{
a character vector of the names of aesthetics permitted by this
geometry object (including alternative localizations of names and base R
equivalent names also accepted by
\href{[https://github.com/tidyverse/ggplot2}{ggplot2}). If \code{geom} is
\code{NULL}, returns all of
\href{[https://github.com/tidyverse/ggplot2}{ggplot2})'s default allowed
aesthetics.
}
\description{
All \href{[https://github.com/tidyverse/ggplot2}{ggplot2}
\code{\link[ggplot2]{Geom}} objects have a set of allowable aesthetics that
are used for generating the layer of the plot. This function returns those
aesthetics, as well as any alternative localizations and base R equivalent
names which can all be used as inputs to specify aesthetics to a ggplot
object.
}
\examples{
library(ggplot2)
ggpackets:::allowed_aesthetics(GeomBar)

}
\seealso{
\code{\link[=filter_aesthetics]{filter_aesthetics()}}, \code{\link[=remove_aesthetics]{remove_aesthetics()}},
\code{\link[=flatten_aesthetics_to_group]{flatten_aesthetics_to_group()}}, \code{\link[=add_eqv_aes]{add_eqv_aes()}}, \code{\link[=split_aes_from_dots]{split_aes_from_dots()}},
\code{\link[=is_uneval]{is_uneval()}}
}
\keyword{aes}
\keyword{aesthetics}
