% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics-handling.r
\name{allowed_params}
\alias{allowed_params}
\title{Get allowed paramters for ggplot objects}
\usage{
allowed_params(geom = NULL)
}
\arguments{
\item{geom}{A Geom (or Stat) object}
}
\value{
the allowed paramters for the given object
}
\description{
Get allowed paramters for ggplot objects
}
