% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_df.r
\name{append_calldf}
\alias{append_calldf}
\title{Append dataframe to another after row index}
\usage{
append_calldf(x, values, after = nrow(x))
}
\arguments{
\item{x}{an call_df to be appended to}

\item{values}{the call_df, whose arguments will be inserted}

\item{after}{the rowindex at which to insert the dataframe}
}
\value{
a combined call_df with call_df \code{values}'s args inserted into
call_df \code{x} after row \code{after}
}
\description{
Append dataframe to another after row index
}
