% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_ggpk.r
\name{as.expression.function}
\alias{as.expression.function}
\title{Coersion from function to expression}
\usage{
\method{as.expression}{function}(x, ...)
}
\arguments{
\item{x}{the function handle of the function to coerce}

\item{...}{additional arguments will be ignored}
}
\value{
the expression representation of the function body
}
\description{
Coersion from function to expression
}
