% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpacket.R
\name{as.list,ggpacket-method}
\alias{as.list,ggpacket-method}
\title{Convert a ggpacket to a list of ggcalls}
\usage{
\S4method{as.list}{ggpacket}(x, ...)
}
\arguments{
\item{x}{A \code{ggpacket} object}

\item{...}{Additional arguments unused}
}
\description{
Convert a ggpacket to a list of ggcalls
}
