% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcall.R
\name{as_gg_call}
\alias{as_gg_call}
\title{Convert an expression into a call as a list of quosure components}
\usage{
as_gg_call(x, which = -3L)
}
\description{
Convert an expression into a call as a list of quosure components
}
