% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_df.r
\name{bind_call}
\alias{bind_call}
\title{Bind a call to a calldf}
\usage{
bind_call(x, .call, name, envir = parent.frame(), match_args = FALSE)
}
\arguments{
\item{x}{the calldf object to modify}

\item{.call}{the call to bind to the calldf}

\item{name}{a character vector to use for the call name}

\item{envir}{the environment in which the call is to be executed}

\item{match_args}{whether unnamed arguments should be matched against named
call formals}
}
\value{
the calldf object with the necessary call data added
}
\description{
If the call produces a \code{ggplot2} \code{ggproto} object, the \code{geom}
and \code{stat} of the \code{ggproto} are cached in the calldf object for use
in ggplot construction.
}
