% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_df.r
\name{build_proto}
\alias{build_proto}
\title{Build a minimal ggproto object from desired arguments}
\usage{
build_proto(.call, args)
}
\arguments{
\item{.call}{the function to use for ggproto construction}

\item{args}{arguments to use when calling the function}
}
\value{
a minimal ggproto object using only the geom and stat arguments of
args, safely returning an empty list with geom and stat values if an error
is encountered
}
\description{
Build a minimal ggproto object from desired arguments
}
