% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_df.r
\docType{methods}
\name{as.list,call_df-method}
\alias{as.list,call_df-method}
\title{Caste call_df to list of arguments, optionally filtering duplicates}
\usage{
\S4method{as.list}{call_df}(x, duplicates = FALSE)
}
\arguments{
\item{x}{call_df object}

\item{duplicates}{whether to allow duplicate arguments in list}
}
\description{
Caste call_df to list of arguments, optionally filtering duplicates
}
