% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_df.r
\docType{class}
\name{call_df-class}
\alias{call_df-class}
\alias{call_df}
\title{A class to contain arguments for lazy evaluation of calls in a ggplot
construction}
\description{
A class to contain arguments for lazy evaluation of calls in a ggplot
construction
}
\section{Slots}{

\describe{
\item{\code{args}}{a tibble containing columns 'name', 'val', and 'source' for
tracking the provenance of call arguments, handling of multiple arguments
and communicating shadowing of duplicate arguments.}

\item{\code{call}}{the call to evaluate during plot construction}

\item{\code{geom}}{the resulting geom of the call, stored upon binding to aid in
argument handling}

\item{\code{stat}}{the resulting stat object of the call, stored upon binding to aid
in argument handling.}

\item{\code{auto_remove_aes}}{whether to automatically remove aesthetic arguments
that aren't relevant to the given geom and stat objects.}
}}

