% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_df.r
\docType{methods}
\name{==,call_df,call_df-method}
\alias{==,call_df,call_df-method}
\title{call_df equivalence}
\usage{
\S4method{==}{call_df,call_df}(e1, e2)
}
\arguments{
\item{e1}{lhs}

\item{e2}{rhs}
}
\description{
call_df equivalence
}
