% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_df.r
\docType{methods}
\name{[,call_df-method}
\alias{[,call_df-method}
\alias{[[,call_df-method}
\alias{$,call_df-method}
\title{Indexing into argument tibble rows by vector}
\usage{
\S4method{[}{call_df}(x, i, j, subset = c("named", "unnamed"))

\S4method{[[}{call_df}(x, i)

\S4method{$}{call_df}(x, name)
}
\arguments{
\item{x}{call_df object to index into}

\item{i}{row index of element to select}

\item{j}{column index of element to select}

\item{subset}{criteria to subset resulting calldf by}

\item{name}{index}
}
\description{
Indexing into argument tibble rows by vector

Indexing into arguments by name

Indexing into arguments by name
}
