% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_df.r
\docType{methods}
\name{show,call_df-method}
\alias{show,call_df-method}
\title{Show call_df contents}
\usage{
\S4method{show}{call_df}(object)
}
\arguments{
\item{object}{call_df object}
}
\description{
Show call_df contents
}
