% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_arguments.R
\name{collapse_data}
\alias{collapse_data}
\title{Collapse data arguments}
\usage{
collapse_data(args)
}
\arguments{
\item{args}{A list of arguments. If multiple elements are named "data"
they will be collapsed into a single mapping argument, appropriately
updating datasets and applying functions as needed.}
}
\description{
Reduce data arguments using \code{update_data}
}
