% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_arguments.R
\name{collapse_mappings}
\alias{collapse_mappings}
\title{Collapse aesthetic mappings arguments}
\usage{
collapse_mappings(args)
}
\arguments{
\item{args}{A list of arguments. If multiple elements are named "mapping"
they will be collapsed into a single mapping argument, appropriately
updating aesthetic mappings.}
}
\description{
Reduce aesthetic mapping arguments using \code{update_mapping}
}
