% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyeval.r
\name{dequo}
\alias{dequo}
\title{remove quoting of quoted item, returning only character name}
\usage{
dequo(q)
}
\arguments{
\item{q}{argument to de-quote}
}
\description{
remove quoting of quoted item, returning only character name
}
