% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyeval.r
\name{dequos}
\alias{dequos}
\title{remove quoting of quoted list items, returning only character names}
\usage{
dequos(...)
}
\arguments{
\item{...}{arguments to de-quote}
}
\description{
remove quoting of quoted list items, returning only character names
}
