% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_df.r
\name{do_calldf}
\alias{do_calldf}
\title{do.call equivalent for calldf objects}
\usage{
do_calldf(what, envir = parent.frame())
}
\arguments{
\item{what}{the calldf object to evaluate}

\item{envir}{the environment in which to evaluate the call}
}
\description{
do.call equivalent for calldf objects
}
