% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_ggpk.r
\name{document_ggpk}
\alias{document_ggpk}
\title{Generate roxygen ellipses param documentation from function code}
\usage{
document_ggpk(f, header = TRUE, ..., prefix = "#' ",
  dots_as_ellipses = FALSE)
}
\arguments{
\item{f}{function to generate ellipses documentation for}

\item{header}{whether to include a brief intro, adding context to arg list}

\item{...}{additional arguments used for substituting unknown variables}

\item{prefix}{prefix to insert before lines of documentation}

\item{dots_as_ellipses}{build documentation for arguments passed via dots
argument}
}
\value{
a block of roxygen text for documenting the ellipses parameter
}
\description{
Generate roxygen ellipses param documentation from function code
}
\examples{
document_ggpk(ggpk_ty)

}
