% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_aesthetics.R
\docType{data}
\name{.all_aesthetics}
\alias{.all_aesthetics}
\title{Extracted .all_aesthetics from internal ggplot2 with hardcoded fallback}
\format{
An object of class \code{character} of length 41.
}
\usage{
.all_aesthetics
}
\description{
Extracted .all_aesthetics from internal ggplot2 with hardcoded fallback
}
\keyword{datasets}
