% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_ext.R
\name{.get_ggplot2_internal_plus_gg}
\alias{.get_ggplot2_internal_plus_gg}
\title{Retrieve ggplot2 internal gg addition method}
\usage{
.get_ggplot2_internal_plus_gg()
}
\value{
ggplot2's unexported \code{+.gg} handler
}
\description{
Retrieve ggplot2 internal gg addition method
}
\keyword{internal}
