% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_ext.R
\name{.plus_gg}
\alias{.plus_gg}
\title{ggplot2 internal gg addition method}
\usage{
.plus_gg(e1, e2)
}
\arguments{
\item{e1}{Addition lhs}

\item{e2}{Addition rhs}
}
\description{
ggplot2 internal gg addition method
}
