% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpacket.r
\name{drop_envs}
\alias{drop_envs}
\title{used to help with unit testing}
\usage{
drop_envs(ggpk)
}
\arguments{
\item{ggpk}{ggpacket object to strip of envir slot value}
}
\description{
used to help with unit testing
}
