% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_df.r
\name{exclude_calldf_args}
\alias{exclude_calldf_args}
\title{Remove calldf arguments by name}
\usage{
exclude_calldf_args(calldf, exclude)
}
\arguments{
\item{calldf}{the calldf object to modify}

\item{exclude}{a vector of argument names to remove from the calldf arugment
tibble}
}
\value{
the modified calldf
}
\description{
Remove calldf arguments by name
}
