% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_arguments.R
\name{expand_dots}
\alias{expand_dots}
\title{Expand dot arguments into named arguments}
\usage{
expand_dots(expr, envir = parent.frame(2L))
}
\arguments{
\item{expr}{An expression whose arguments need to be expanded.}

\item{envir}{An environment in which to evaluate expansion.}
}
\description{
Expand dot arguments into named arguments
}
