% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics-handling.r
\name{filter_aesthetics}
\alias{filter_aesthetics}
\title{Filter aesthetics for Geom}
\usage{
filter_aesthetics(geom, mapping)
}
\arguments{
\item{geom}{a ggplot2 \code{\link[ggplot2]{Geom}} object}

\item{mapping}{a ggplot2 aesthetic mapping}
}
\value{
the provided aesthetic mapping, filtered by accepted aesthetics for
the given \code{\link[ggplot2]{Geom}}
}
\description{
Within ggpackets, often times many aesthetics will be passed to all packed
layers without regard for whether or not their applicable. To accommodate
this, it's often necessary to first filter down to only the applicable
aesthetics for a given \code{\link[ggplot2]{Geom}} object to avoid erroneous
warning messages.
}
\examples{
library(ggplot2)
filter_aesthetics(GeomBar, aes(x = test1, y = test2))
  
}
\seealso{
\code{\link[=allowed_aesthetics]{allowed_aesthetics()}}, \code{\link[=filter_aesthetics]{filter_aesthetics()}},
\code{\link[=remove_aesthetics]{remove_aesthetics()}}, \code{\link[=flatten_aesthetics_to_group]{flatten_aesthetics_to_group()}}, \code{\link[=add_eqv_aes]{add_eqv_aes()}},
\code{\link[=split_aes_from_dots]{split_aes_from_dots()}}, \code{\link[=is_uneval]{is_uneval()}}
}
\keyword{aes}
\keyword{aesthetics}
