% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics-handling.r
\name{filter_args}
\alias{filter_args}
\title{Filter arguments for given Geom and Stat objects}
\usage{
filter_args(call, geom, stat, args)
}
\arguments{
\item{call}{a function call whose formals should be used to filter args}

\item{geom}{Geom object for which argument should be filtered}

\item{stat}{Stat object for which argument should be filtered}

\item{args}{argument list to be filtered}
}
\description{
Filter arguments for given Geom and Stat objects
}
