% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_arguments.R
\name{filter_by_ggcall_ids}
\alias{filter_by_ggcall_ids}
\title{Filter a named list by ids}
\usage{
filter_by_ggcall_ids(x, call_ids, all_ids)
}
\arguments{
\item{x}{A named list to filter.}

\item{call_ids}{A character vector of valid ids.}

\item{all_ids}{A character vector of all possible ids.}
}
\description{
Filter a named list (often an aesthetic mapping or argument list) for only
unprefixed elements, elements which don't match any of \code{all_ids} and
those which match one of \code{call_ids}. For those that match, remove the
id prefix.
}
