% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_ggpk.R
\name{find_calls}
\alias{find_calls}
\title{Find calls to specified function name}
\usage{
find_calls(x, call_name = NULL)
}
\arguments{
\item{x}{expression to parse}

\item{call_name}{call name to search for}
}
\value{
calls introduced in expression
}
\description{
Find calls to specified function name
}
