% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpack.r
\name{flatten_aes_to_mapping}
\alias{flatten_aes_to_mapping}
\title{Move items in args to 'mapping' key if found in aes_list}
\usage{
flatten_aes_to_mapping(args, aes_list = add_eqv_aes(.all_aesthetics),
  filter_mapping = FALSE, envir = parent.frame())
}
\arguments{
\item{args}{a list of arguments to parse for aesthetic arguments}

\item{aes_list}{a list of aesthetic names to consider}

\item{filter_mapping}{whether mapping key should be filtered for only those
in the aes_list terms}

\item{envir}{environment in which to attempt evaluation of mapping term}
}
\value{
a list of arguments with aesthetics matching values in
\code{aes_list} flattened into a list stored within the mapping value
}
\description{
Move items in args to 'mapping' key if found in aes_list
}
