% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics-handling.r
\name{flatten_aesthetics_to_group}
\alias{flatten_aesthetics_to_group}
\title{Flatten aesthetics into group}
\usage{
flatten_aesthetics_to_group(mapping, ...)
}
\arguments{
\item{mapping}{a \href{[https://github.com/tidyverse/ggplot2}{ggplot2}
aesthetic mapping}

\item{...}{aesthetic handles to flatten into the group aesthetic as
interaction terms}
}
\value{
the mapping aesthetics with specified aesthetics stored additionally
as interaction terms in the group aesthetics with any preexisting group
aesthetics
}
\description{
Occasionally, it may be necessary to enforce a group aesthetic from a
selection of other aesthetics that are being passed to a packed layer. In
this case, those aesthetics can be passed and preserved as an interaction
term in the group aesthetic.
}
\details{
For example, if one would want to produce a line plot with an errorbar at
every distinct point along the x-axis and use the linetype aesthetic to group
the lines of the line plot, this will be problematic because it may also
affect the linetype of the errorbars. To accommodate this, the linetype might
be flattened into the group aesthetic, allowing for the linetype of the
errorbars to be overridden without affecting the data groups.
}
\note{
the \code{x} and \code{y} aesthetics will not be pulled into the group aesthetic
}
\examples{
library(ggplot2)
my.aes <- aes(x = mpg, y = wt, color = carb, fill = vs)
flatten_aesthetics_to_group(my.aes, 'color', 'fill')
  
}
\seealso{
\code{\link[=allowed_aesthetics]{allowed_aesthetics()}}, \code{\link[=filter_aesthetics]{filter_aesthetics()}},
\code{\link[=remove_aesthetics]{remove_aesthetics()}}, \code{\link[=flatten_aesthetics_to_group]{flatten_aesthetics_to_group()}}, \code{\link[=add_eqv_aes]{add_eqv_aes()}},
\code{\link[=split_aes_from_dots]{split_aes_from_dots()}}, \code{\link[=is_uneval]{is_uneval()}}
}
\keyword{aes}
\keyword{aesthetics}
