% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpack.r
\name{ggpack}
\alias{ggpack}
\title{Handling of argument passing to ggplots}
\usage{
ggpack(.call = NULL, ..., id = NULL, dots = NULL,
  auto_remove_aes = is.null(id) || any(sapply(id, is.null)),
  null_empty = FALSE, envir = parent.frame())
}
\arguments{
\item{.call}{the \code{\link[ggplot2]{ggplot}} geom function to invoke}

\item{...}{additional arguments to use when calling the provided function}

\item{id}{the prefix string to subset arguments by}

\item{dots}{arguments to subset}

\item{auto_remove_aes}{logical indicating whether aesthetic mappings that}

\item{null_empty}{logical indicating whether \code{NULL} should be returned
if no arguments are passed to call}

\item{envir}{frame in which non-aesthetic arguments should be evaluated}
}
\value{
a call to the specified function with arguments subset for only those
which match the specified prefix
}
\description{
A helper function to wrap ggplot calls, allowing for passing of a shared list
of arguments prefixed by an id string. For example, a call to a custom ggplot
template may allow for arguments 'line.color' and 'bar.color' to specify the
line and bar colors separately. These arguments are parsed appropriately and
passed to the appropriate sub-function.
}
\note{
Ellipses arguments and the arguments passed through the \code{dots} args
will overwrite previous arguments of the same name.
}
\examples{
library(ggplot2)

# wrap layer calls in a ggpack to package them
ggpack(geom_bar(na.rm = TRUE))

# or more explicitly define layer properties
ggpack(geom_bar,              # layer function to bundle
  id = 'bar',                 # name of layer - used for filter listed args
  fill = 'red',               # ...'s before 'args' used as defaults
  dots = list(fill = 'blue'), # named 'args' param will overwrite prior ...'s
  fill = 'green')             # ...'s after 'args' will overwrite any others

}
