% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_df.r, R/ggpacked_layer.r
\docType{methods}
\name{+,call_df,call_df-method}
\alias{+,call_df,call_df-method}
\alias{+,NULL,ggpacked_layer-method}
\alias{+,ggpacked_layer,ggpacked_layer-method}
\alias{+,gg,ggpacked_layer-method}
\title{Sum of two ggpacked layers to return ggpacket object}
\usage{
\S4method{+}{call_df,call_df}(e1, e2)

\S4method{+}{`NULL`,ggpacked_layer}(e1, e2)

\S4method{+}{ggpacked_layer,ggpacked_layer}(e1, e2)

\S4method{+}{gg,ggpacked_layer}(e1, e2)
}
\arguments{
\item{e1}{left side argument for addition}

\item{e2}{right side argument for addition}
}
\description{
Sum of two ggpacked layers to return ggpacket object

Add NULL to a ggpacket to return the ggpacket

Sum of two ggpacked layers to return ggpacket object

Primitive methods for adding ggpacked layers to ggplot during construction
}
