% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpacked_layer.r
\docType{class}
\name{ggpacked_layer-class}
\alias{ggpacked_layer-class}
\alias{ggpacked_layer}
\title{A class for storing ggpack call contents for lazy evaluation}
\description{
A class for storing ggpack call contents for lazy evaluation
}
\section{Slots}{

\describe{
\item{\code{ggcall}}{a constructor function to produce a ggproto layer object}

\item{\code{ggargs}}{the arguments to use when calling the constructor function}
}}

