% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpacked_layer.r
\docType{methods}
\name{==,ggpacked_layer,ggpacked_layer-method}
\alias{==,ggpacked_layer,ggpacked_layer-method}
\title{ggpacket_layer equivalence}
\usage{
\S4method{==}{ggpacked_layer,ggpacked_layer}(e1, e2)
}
\arguments{
\item{e1}{lhs}

\item{e2}{rhs}
}
\description{
ggpacket_layer equivalence
}
