% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpacked_layer.r, R/ggpacket.r
\docType{methods}
\name{[,ggpacked_layer-method}
\alias{[,ggpacked_layer-method}
\alias{[[,ggpacked_layer-method}
\alias{$,ggpacked_layer-method}
\alias{$,ggpacket-method}
\title{Index into layer arguments with vector}
\usage{
\S4method{[}{ggpacked_layer}(x, i)

\S4method{[[}{ggpacked_layer}(x, i)

\S4method{$}{ggpacked_layer}(x, name)

\S4method{$}{ggpacket}(x, name)
}
\arguments{
\item{x}{ggpacked_layer object}

\item{i}{index}

\item{name}{index}
}
\description{
Index into layer arguments with vector

Index into layer arguments by numeric or character

Index into layer arguments by name
}
