% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpacket.r
\docType{methods}
\name{+,ggpacket,ANY-method}
\alias{+,ggpacket,ANY-method}
\alias{+,NULL,ggpacket-method}
\alias{+,gg,ggpacket-method}
\title{Primitive methods for adding ggpackets to various ggplot objects}
\usage{
\S4method{+}{ggpacket,ANY}(e1, e2)

\S4method{+}{`NULL`,ggpacket}(e1, e2)

\S4method{+}{gg,ggpacket}(e1, e2)
}
\arguments{
\item{e1}{left side argument for addition}

\item{e2}{right side argument for addition}
}
\description{
Primitive methods for adding ggpackets to various ggplot objects

Add NULL to a ggpacket to return the ggpacket

Combine ggpacket contents with gg call to incorporate seamlessly with ggplot calls
}
