% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpacket.r
\docType{methods}
\name{==,ggpacket,ggpacket-method}
\alias{==,ggpacket,ggpacket-method}
\title{ggpacket equivalence}
\usage{
\S4method{==}{ggpacket,ggpacket}(e1, e2)
}
\arguments{
\item{e1}{lhs}

\item{e2}{rhs}
}
\description{
ggpacket equivalence
}
