% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpacket.r
\docType{methods}
\name{[,ggpacket-method}
\alias{[,ggpacket-method}
\alias{[[,ggpacket-method}
\title{Overload [ generic to index into ggcalls list}
\usage{
\S4method{[}{ggpacket}(x, i)

\S4method{[[}{ggpacket}(x, i)
}
\arguments{
\item{x}{ggpacket object}

\item{i}{index of the ggcall in the ggpacket object}
}
\description{
Overload [ generic to index into ggcalls list

Overload [[ generic to index into ggcalls list
}
