% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpacket.R, R/ggpacket_show.R
\name{+,ggpacket,ANY-method}
\alias{+,ggpacket,ANY-method}
\alias{show,ggpacket-method}
\title{Add a ggpacket object to another, arbitrary object}
\usage{
\S4method{+}{ggpacket,ANY}(e1, e2)

\S4method{show}{ggpacket}(object)
}
\arguments{
\item{e1}{A ggpacket object.}

\item{e2}{Any object.}

\item{object}{A ggpacket object to show.}
}
\description{
Add a ggpacket object to another, arbitrary object

Display contents of a ggpacket
}
