% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpackets.r
\docType{package}
\name{ggpackets-package}
\alias{ggpackets}
\alias{ggpackets-package}
\title{ggpackets}
\description{
A package for easier compositing and reusing of common ggplots.
}
\details{
\code{ggpackets} aims to make composite \code{ggplot2} functions:
\enumerate{
\item more reusable
\item easier to develop multilayer templates
\item functionally interoperable with ggplot2
\item syntactically similar to ggplot2
}

\code{ggpackets} also tries to expose functions that allow for use of the package
at varying levels of syntactic intricacy. At its most clean and
straightforward extreme, \code{ggpackets} exposes functions which appear
syntactically identical to \code{ggplot}. As more is demanded from the provided
functions, the package aims to accommodate those needs gradually introducing
additional parameters.
}
\author{
\strong{Maintainer}: Doug Kelkhoff \email{kelkhoff.douglas@gene.com}

}
\keyword{internal}
