% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpks.r
\name{ggpk_decorators}
\alias{ggpk_decorators}
\title{Wrapper for common decorators to package}
\usage{
ggpk_decorators(...)
}
\arguments{
\item{...}{arguments to be passed to any of the bundled decorator plot
functions, blank, xlab, ylab, labs, theme or ggtitle prefaced by any of
those strings ('labs.title').}
}
\value{
a ggpacket object containing a collection of decorating plot
functions.
}
\description{
Wrapper for common decorators to package
}
\examples{
library(ggplot2)
ggplot(mtcars) + 
  aes(x = carb) + 
  geom_bar() + 
  ggpk_decorators(
    xlab = 'Number of Carburetors',
    ylab = 'Count of Cars',
    labs.title = 'Distribution of Carburetors',
    labs.caption = '* taken from 1974 Motor Trends US')

}
