% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpks.r
\name{ggpk_ty}
\alias{ggpk_ty}
\title{Say thanks!}
\usage{
ggpk_ty(..., xmin = 0, xmax = 1, ymin = 0, ymax = 1)
}
\arguments{
\item{...}{aesthetic arguments to pass}

\item{xmin}{minimum x value to fill with text}

\item{xmax}{maximum x value to fill with text}

\item{ymin}{minimum y value to fill with text}

\item{ymax}{maximum y value to fill with text}
}
\description{
Say thanks!
}
