% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_ext.R
\name{\%+\%}
\alias{\%+\%}
\title{Lazy handler for ggplot addition}
\usage{
e1 \%+\% e2
}
\arguments{
\item{e1}{Addition lhs.}

\item{e2}{Addition rhs.}
}
\value{
A new \code{ggpacket} object with \code{e2} appended as an additional
layer or layers.
}
\description{
Lazy handler for ggplot addition
}
