% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_aesthetics.R
\name{handle_reset_mapping}
\alias{handle_reset_mapping}
\title{Specific handling of ..reset.. aesthetic}
\usage{
handle_reset_mapping(mapping)
}
\arguments{
\item{mapping}{A ggplot2 aesthetic mapping.}
}
\description{
Specific handling of ..reset.. aesthetic
}
