% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpacket.R
\docType{methods}
\name{ggpacket-class}
\alias{ggpacket-class}
\title{A ggpacket object}
\description{
A ggpacket object
}
\section{Slots}{

\describe{
\item{\code{data}}{A dataset or waiver to use within the ggpacket ggcalls.}

\item{\code{mapping}}{A ggplot aesthetic mapping to use within the ggpacket ggcalls.}

\item{\code{dots}}{Arguments which should be passed before prespecified arguments to
each ggcall.}

\item{\code{ggcalls}}{A list of lazily evaluated ggplot layer construction
expressions.}
}}

