% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{if-not-null-else}
\alias{if-not-null-else}
\alias{\%||\%}
\title{if not NULL else}
\usage{
a \%||\% b
}
\arguments{
\item{a}{left hand side of operator}

\item{b}{right hand side of operator}
}
\value{
b if a is null, otherwise a
}
\description{
if not NULL else
}
