% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcall.R
\name{infer_ggcall_id}
\alias{infer_ggcall_id}
\title{Convert ggplot geom layers to friendly names}
\usage{
infer_ggcall_id(x)
}
\arguments{
\item{x}{A function name from which an id should be inferred.}
}
\description{
Convert ggplot geom layers to friendly names
}
