% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpack.r
\name{invalid_uneval_aes_match}
\alias{invalid_uneval_aes_match}
\title{Indices of unevaluated aesthetics not in aes_list}
\usage{
invalid_uneval_aes_match(args, expr = NULL,
  aes_list = add_eqv_aes(.all_aesthetics))
}
\arguments{
\item{args}{a list of arguments which may contain unevaluated values}

\item{expr}{a regular expression exclude unevaluated values with. default is
NULL and will remove all unevaluated values that match aesthetic names}

\item{aes_list}{a list of aesthetic names to consider for removal}
}
\value{
a logical vector, with unevaluated values set as TRUE if the regex
match matches an aesthetic name from aes_list, otherwise FALSE
}
\description{
Indices of unevaluated aesthetics not in aes_list
}
