% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpack.r
\name{last_args}
\alias{last_args}
\title{Filter argument list for only unnamed items and the last instance of each
named item}
\usage{
last_args(args, sources = list(), warn = c(), desc = c(), call = NULL)
}
\arguments{
\item{args}{a list of arguments to filter}

\item{sources}{an optional list of names used to indicate the origin of each
of the arguments}

\item{warn}{an optional character vector of sources for which a warning
should be produced, should an argument from that source be overwritten}

\item{desc}{an optional named character vector of more verbose descriptions
to use for each of the sources. (e.g. \code{c("..." = "ellipses
  arguments")})}

\item{call}{an optional object of type call to include with the warning if a
statment associated with the arguments is desired to be included with the
warning message.}
}
\value{
a named list of arguments with duplicated names omitted, leaving only
the last occurance of each named value
}
\description{
Filter argument list for only unnamed items and the last instance of each
named item
}
