% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_df.r
\docType{methods}
\name{length,call_df-method}
\alias{length,call_df-method}
\title{Overload length method to retrieve number of arguments from calldf}
\usage{
\S4method{length}{call_df}(x)
}
\arguments{
\item{x}{the calldf object to show}
}
\description{
Overload length method to retrieve number of arguments from calldf
}
