% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpacket.R
\name{length,ggpacket-method}
\alias{length,ggpacket-method}
\title{Get the number of ggcalls within a ggpacket}
\usage{
\S4method{length}{ggpacket}(x)
}
\arguments{
\item{x}{A \code{ggpacket} object}
}
\description{
Get the number of ggcalls within a ggpacket
}
