% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpack.r
\name{list_diff}
\alias{list_diff}
\title{Attempt to identify which values were added in the transition from a to b}
\usage{
list_diff(a, b)
}
\arguments{
\item{a}{a list of values}

\item{b}{another list of values}
}
\value{
the values that were added to a to arrive at b
}
\description{
Attempt to identify which values were added in the transition from a to b
}
\note{
This can cause issues if ambiguous terms are in b, making it impossible
to distinguish which arguments were added. For example:

\code{a = c('a', 'b', 'b', 'b', 'a')}
\code{b = c('a', 'b', 'b', 'c', 'b', 'b', 'a')}

makes it impossible to know whether c('b', 'c') was inserted after position
2, or whether c('c', 'b') was inserted after position 3.
}
