% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_df.r
\name{match_calldf}
\alias{match_calldf}
\title{Match arguments in the calldf against the formals of the call function}
\usage{
match_calldf(calldf)
}
\arguments{
\item{calldf}{the calldf object for which arguments should be matched against
call formals}
}
\value{
a calldf object of the same structure, with arugment names provided
for unnamed arguments.
}
\description{
a \code{match.call} equivalent for the calldf object
}
\examples{
# intended for internal ggpackets use only

my_calldf <- call_df(call = print, arglist = list("a"), source = 'ex')
my_calldf
#> function:
#>   ex .. "a"

my_calldf <- ggpackets:::match_calldf(my_calldf)
#> function:
#>   ex .. x = "a"

}
