% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_arguments.R
\name{match_unnamed_args}
\alias{match_unnamed_args}
\title{Match unnamed arguments}
\usage{
match_unnamed_args(f, args, envir = parent.frame())
}
\arguments{
\item{f}{A function to match arguments against.}

\item{args}{A list of arguments to match.}

\item{envir}{An environment in which the function should be matched.}
}
\description{
Similar to match.call, but without evaluating arguments, avoiding possible
syntactic errors that would arise due to ggpackets-specific syntax.
}
