% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpack.r
\name{named_uneval_aes}
\alias{named_uneval_aes}
\title{Indices of any unevaluated aesthetics with a prefix}
\usage{
named_uneval_aes(args, aes_list = add_eqv_aes(.all_aesthetics))
}
\arguments{
\item{args}{a list of arguments which may contain unevaluated values}

\item{aes_list}{a list of aesthetic names to consider for removal}
}
\value{
a logical vector, with unevaluated values set as TRUE if aesthetic
name from aes_list, otherwise FALSE
}
\description{
A prefix comes in the form 'prefix.aesthetic_name'
}
