% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_df.r
\docType{methods}
\name{names,call_df-method}
\alias{names,call_df-method}
\alias{names<-,call_df-method}
\title{Overload names method to print calldf argument names}
\usage{
\S4method{names}{call_df}(x)

\S4method{names}{call_df}(x) <- value
}
\arguments{
\item{x}{the calldf object to show}

\item{value}{the new values to pass to names of call_df}
}
\description{
Overload names method to print calldf argument names

Overload names method to print calldf argument names
}
