% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpacket.R
\name{names,ggpacket-method}
\alias{names,ggpacket-method}
\title{Fetch the ids associated with each ggcall}
\usage{
\S4method{names}{ggpacket}(x)
}
\arguments{
\item{x}{A \code{ggpacket} object}
}
\description{
Fetch the ids associated with each ggcall
}
