% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpacket.r
\docType{methods}
\name{names,ggpacket-method}
\alias{names,ggpacket-method}
\title{Overload names method to print ggpacket ggcall list names}
\usage{
\S4method{names}{ggpacket}(x)
}
\arguments{
\item{x}{the ggpacket object to show}
}
\description{
Overload names method to print ggpacket ggcall list names
}
