% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_arguments.R
\name{only_formals_and_dots}
\alias{only_formals_and_dots}
\title{Filter for only arguments that can be accepted by a given function}
\usage{
only_formals_and_dots(f, args)
}
\arguments{
\item{f}{A function to filter on.}

\item{args}{A list of arguments to filter.}
}
\description{
Filter for only arguments that can be accepted by a given function
}
