% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_ext.R
\name{+.gg}
\alias{+.gg}
\title{Intercept ggplot2 ggproto plus operator}
\usage{
\method{+}{gg}(e1, e2)
}
\arguments{
\item{e1}{An object to add to a ggproto object.}

\item{e2}{A ggproto object to add.}
}
\description{
Intercept ggplot2 ggproto plus operator
}
