% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics-handling.r
\name{remove_aesthetics}
\alias{remove_aesthetics}
\title{Remove specific aesthetics from mapping}
\usage{
remove_aesthetics(mapping, ...)
}
\arguments{
\item{mapping}{aesthetic mapping to use for filtering}

\item{...}{mapping names to filter out}
}
\value{
an aesthetic mapping with filtered mappings removed and group mapping
set as interaction terms of all non axial (non- x, y) terms.
}
\description{
Remove specified aesthetics by name from a
\href{[https://github.com/tidyverse/ggplot2}{ggplot2} aesthetic mapping,
returning the same mapping with the specified aesthetics removed an preserved
as an interaction term in the group aesthetic.
}
\examples{
library(ggplot2)
remove_aesthetics(aes(x = wt, y = mpg, color = carb, fill = wt), 'fill')

}
\seealso{
\code{\link[=allowed_aesthetics]{allowed_aesthetics()}}, \code{\link[=filter_aesthetics]{filter_aesthetics()}},
\code{\link[=remove_aesthetics]{remove_aesthetics()}}, \code{\link[=flatten_aesthetics_to_group]{flatten_aesthetics_to_group()}}, \code{\link[=add_eqv_aes]{add_eqv_aes()}},
\code{\link[=split_aes_from_dots]{split_aes_from_dots()}}, \code{\link[=is_uneval]{is_uneval()}}
}
\keyword{aes}
\keyword{aesthetics}
