% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpack.r
\name{remove_by_prefix}
\alias{remove_by_prefix}
\title{Filter arguments based on a prefix id}
\usage{
remove_by_prefix(id, args, sep = "\\\\.")
}
\arguments{
\item{id}{a string specifying the prefix to be used to filter args}

\item{args}{a list of arguments to be subset by the prefix match}

\item{sep}{a regex joining string between prefix and args.
Defaults to \code{"\\\\."}.}
}
\value{
a list of arguments that originally were prefaced by the
specified prefix, now with that prefix removed.
}
\description{
Filter arguments based on a prefix id
}
