% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics-handling.r
\name{rename_to_ggplot}
\alias{rename_to_ggplot}
\title{Rename non-ggplot2 standard aesthetics to ggplot2 standard}
\usage{
rename_to_ggplot(args)
}
\arguments{
\item{args}{a list of arguments which may or may not contain aesthetics}
}
\value{
the list with aesthetic names remapped to ggplot equivalents
}
\description{
Rename non-ggplot2 standard aesthetics to ggplot2 standard
}
