% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_df.r
\name{replace_reserved_aesthetic_references}
\alias{replace_reserved_aesthetic_references}
\title{Filter reserved aesthetics of the form ..<aes>.. from values}
\usage{
replace_reserved_aesthetic_references(args, re = "^\\\\.\\\\.(.+)\\\\.\\\\.$")
}
\arguments{
\item{args}{named argument list potentially containing a reserved value}

\item{re}{the regular expression to match deparsed symbols against}
}
\value{
the argument list with reserved values replaced with the value from
the name of the element if one occurs before the reserved value or removed
otherwise
}
\description{
Filter reserved aesthetics of the form ..<aes>.. from values
}
