% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpacket.R
\name{required_aesthetics}
\alias{required_aesthetics}
\title{Check a ggpacket object for required aesthetic arguments}
\usage{
required_aesthetics(x)
}
\arguments{
\item{x}{A \code{ggpacket} object or related \code{ggplot} component}
}
\description{
Check a ggpacket object for required aesthetic arguments
}
