% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{safecrayon}
\alias{safecrayon}
\title{Safe use of package:crayon exports with fallback to paste(...)}
\usage{
safecrayon(fname, ...)
}
\arguments{
\item{fname}{package:crayon function name to call}

\item{...}{one or more \R objects, to be converted to character vectors.}
}
\description{
Safe use of package:crayon exports with fallback to paste(...)
}
