% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpack.r
\name{safedequos}
\alias{safedequos}
\title{Capture all ellipses arguments using rlang::quos() with substitute() fallback}
\usage{
safedequos(...)
}
\arguments{
\item{...}{arguments from a parent function to extract}
}
\value{
either an rlang list of quosures if rlang is installed or a named
list of values otherwise
}
\description{
Capture all ellipses arguments using rlang::quos() with substitute() fallback
}
