% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{safenchar}
\alias{safenchar}
\title{Safe use of crayon's col_nchar() with base nchar() fallback}
\usage{
safenchar(...)
}
\arguments{
\item{...}{passed to \code{nchar}}
}
\description{
Safe use of crayon's col_nchar() with base nchar() fallback
}
