% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{safesubstr}
\alias{safesubstr}
\title{Safe use of crayon's col_substr() with base substring() fallback}
\usage{
safesubstr(...)
}
\arguments{
\item{...}{passed to \code{substring}}
}
\description{
Safe use of crayon's col_substr() with base substring() fallback
}
