% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpacket.R
\name{self}
\alias{self}
\title{Returning the calling object from within a function}
\usage{
self(which = -1L)
}
\arguments{
\item{which}{A relative environment offset in which to search for an object
with a name of the calling expression.}
}
\description{
Used for retreiving an S4 object being called as though it is a function
}
