% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpacked_layer.r
\docType{methods}
\name{show,ggpacked_layer-method}
\alias{show,ggpacked_layer-method}
\title{Overload show method to show ggpacket}
\usage{
\S4method{show}{ggpacked_layer}(object)
}
\arguments{
\item{object}{the ggpacket object to show}
}
\description{
Overload show method to show ggpacket
}
