% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_arguments.R
\name{smart_swap_mapping_data}
\alias{smart_swap_mapping_data}
\title{Mimic ggplot2 behavior of intelligently interpretting first layer argument}
\usage{
smart_swap_mapping_data(args)
}
\arguments{
\item{args}{a list of arguments to interpet}
}
\description{
Like ggplot, if the first argument doesn't appear to be an aesthetic
mapping, despite mappings being appropriately passed in the first argument,
swap the first two arguments.
}
