% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics-handling.r
\name{split_aes_from_dots}
\alias{split_aes_from_dots}
\title{Filter arguments into aesthetic and non-aesthetic}
\usage{
split_aes_from_dots(..., geom = NULL)
}
\arguments{
\item{...}{arguments containing aesthetics mapping}

\item{geom}{if specified, a ggplot2 \code{\link[ggplot2]{Geom}} object can be
passed to filter aesthetics for only a specified
\code{\link[ggplot2]{Geom}}.}
}
\value{
a list containing two named elements: 'aes' which contains the
aesthetic mappings (of the \code{\link[ggplot2]{Geom}} object if specified)
as a pairlist with class \code{uneval} akin to the result of an
\code{\link[ggplot2]{aes}} call, and 'not_aes' which contains a pairlist of
all other arguments
}
\description{
To accommodate passing of aesthetics as equivalent ellipses arguments without
the need for wrapping in an \code{\link[ggplot2]{aes}} call, it's often
necessary to filter out any arguments which share the name of an acceptable
aesthetic argument for the given \code{\link[ggplot2]{Geom}}. In this case,
arguments are split into two separate lists and returned within a list.
}
\examples{
library(ggplot2)
split_aes_from_dots(x = aes_x, my.arg.2 = 'my arg', color = aes_color, 
  fill = 'aes_fill', my.arg.3 = list(1, 2, 3), geom = GeomBar)
  
}
\seealso{
\code{\link[=allowed_aesthetics]{allowed_aesthetics()}}, \code{\link[=filter_aesthetics]{filter_aesthetics()}},
\code{\link[=remove_aesthetics]{remove_aesthetics()}}, \code{\link[=flatten_aesthetics_to_group]{flatten_aesthetics_to_group()}}, \code{\link[=add_eqv_aes]{add_eqv_aes()}},
\code{\link[=split_aes_from_dots]{split_aes_from_dots()}}, \code{\link[=is_uneval]{is_uneval()}}
}
\keyword{aes}
\keyword{aesthetics}
