% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{str_format_list}
\alias{str_format_list}
\title{Format a list of items as it would be in a sentance}
\usage{
str_format_list(l = NULL, oxford = FALSE)
}
\arguments{
\item{l}{a character vector list of strings to format}

\item{oxford}{whether to include an oxford comma}
}
\value{
the items of the input list, formatted as an articulation of a list
in speech (e.g. 'a, b, c and d')
}
\description{
Format a list of items as it would be in a sentance
}
\examples{
str_format_list(c('a', 'b', 'c'))
## [1] "a, b and c"
}
