% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpacket.R
\name{[,ggpacket-method}
\alias{[,ggpacket-method}
\title{Index into a ggpacket object}
\usage{
\S4method{[}{ggpacket}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{A \code{ggpacket} object.}

\item{i}{A \code{character} or \code{numeric} vector for indexing.}

\item{j}{Unused.}

\item{...}{Unused.}

\item{drop}{Unused.}
}
\description{
Index into a ggpacket object
}
