% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpacket.R, R/ggpacket_index.R
\name{[[.ggpacket}
\alias{[[.ggpacket}
\title{Subset a ggpacket for a selection of ggcalls}
\usage{
\method{[[}{ggpacket}(x, i, ...)

\method{[[}{ggpacket}(x, i, ...)
}
\arguments{
\item{x}{A ggpacket to subset.}

\item{i}{A vector upon which to subset the ggpacket ggcalls.}

\item{...}{Additional arguments unused.}
}
\description{
Subset a ggpacket for a selection of ggcalls

Subset a ggpacket for a selection of ggcalls
}
