% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_df.r
\name{substitute_calldf_args}
\alias{substitute_calldf_args}
\title{Substitute calldf named arguments with new argument tibble}
\usage{
substitute_calldf_args(calldf, ...)
}
\arguments{
\item{calldf}{the calldf object to substitute}

\item{...}{named arguments where the name specifies the argument to replace
and the value is a calldf to insert in that argument's place}
}
\value{
the modified calldf
}
\description{
Substitute calldf named arguments with new argument tibble
}
