% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_aesthetics.R
\name{substitute_ggcall_dot_aes}
\alias{substitute_ggcall_dot_aes}
\title{Substitute a ggcall's dot aesthetics with their redirected values}
\usage{
substitute_ggcall_dot_aes(mapping, ggcall, envir = parent.frame())
}
\arguments{
\item{mapping}{A ggplot2 aesthetic mapping.}

\item{ggcall}{A ggcall list of expressions.}

\item{envir}{An environment in which the dot aesthetics should be evaluated.}
}
\description{
Substitute a ggcall's dot aesthetics with their redirected values
}
