% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_aesthetics.R
\name{substitute_quote}
\alias{substitute_quote}
\title{Substitute a quoted expression in a given environmment}
\usage{
substitute_quote(q, env = parent.frame())
}
\arguments{
\item{q}{A quote to evaluate.}

\item{env}{An environment in which the quote should be evaluated.}
}
\description{
Substitute a quoted expression in a given environmment
}
