% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{tsnames}
\alias{tsnames}
\title{Type-safe 'names()' call, always returning a character vector
(potentially containing NA's where list elements are unnamed)}
\usage{
tsnames(l, fill = NA)
}
\arguments{
\item{l}{a list from which names should be pulled}

\item{fill}{a value to use for padding name values to length}
}
\description{
Type-safe 'names()' call, always returning a character vector
(potentially containing NA's where list elements are unnamed)
}
