% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpacket.R
\name{update_data}
\alias{update_data}
\title{Reduce data parameters, iteratively applying functions or masking datasets}
\usage{
update_data(...)
}
\description{
Reduce data parameters, iteratively applying functions or masking datasets
}
