% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpacket.R
\name{update_mapping}
\alias{update_mapping}
\title{Reduce a list of mappings, iteratively routing aesthetics}
\usage{
update_mapping(...)
}
\description{
Reduce a list of mappings, iteratively routing aesthetics
}
