% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_arguments.R
\name{with_ignore_unknown_params}
\alias{with_ignore_unknown_params}
\title{Evaluate an expression, ignoring warnings about unknown parameters}
\usage{
with_ignore_unknown_params(expr, envir = parent.frame())
}
\arguments{
\item{expr}{An expression to evaluate.}

\item{envir}{An environment to evaluate the given expression in.}
}
\description{
Evaluate an expression, ignoring warnings about unknown parameters
}
