% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_aesthetics.R
\name{.all_aesthetics}
\alias{.all_aesthetics}
\title{Extracted .all_aesthetics from internal ggplot2 with hardcoded fallback}
\usage{
.all_aesthetics()
}
\description{
Extracted .all_aesthetics from internal ggplot2 with hardcoded fallback
}
