% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpacket.R, R/ggpacket_show.R
\docType{methods}
\name{ggpacket-class}
\alias{ggpacket-class}
\alias{+,ggpacket,ANY-method}
\alias{show,ggpacket-method}
\alias{show,}
\alias{ggpacket-method}
\title{A ggpacket object}
\usage{
\S4method{+}{ggpacket,ANY}(e1, e2)

\S4method{show}{ggpacket}(object)
}
\arguments{
\item{e1}{A ggpacket object.}

\item{e2}{Any object.}

\item{object}{A ggpacket object to show.}
}
\description{
A ggpacket object

Add a ggpacket object to another, arbitrary object

Display contents of a ggpacket
}
\section{Slots}{

\describe{
\item{\code{data}}{A dataset or waiver to use within the ggpacket ggcalls.}

\item{\code{mapping}}{A ggplot aesthetic mapping to use within the ggpacket ggcalls.}

\item{\code{dots}}{Arguments which should be passed before prespecified arguments to
each ggcall.}

\item{\code{ggcalls}}{A list of lazily evaluated ggplot layer construction
expressions.}
}}

