% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpacket.R
\name{ggpacket_call}
\alias{ggpacket_call}
\title{The function used when a \code{ggpacket} is called as a function}
\usage{
ggpacket_call(data = NULL, mapping = NULL, ...)
}
\arguments{
\item{data}{The data argument to use for all layers within the
\code{ggpacket}}

\item{mapping}{The aesthetic mapping to use as a basis for all layers within
the \code{ggpacket}. Layer-specific mappings will update the value of the
mapping, and assigning \code{..reset..} to any aesthetic field will return
the propegated aesthetic mapping to a default, unmapped state for that
layer.}

\item{...}{additional arguments passed to all bundled \code{ggplot} layers,
and will be overwritten by layer-specific arguments if provided.}
}
\value{
A new \code{ggpacket} object with the new defaults applied
}
\description{
The function used when a \code{ggpacket} is called as a function
}
