% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpacket.R
\name{ggpacket_plus_ANY}
\alias{ggpacket_plus_ANY}
\title{Swallow calls when a ggpacket is added to any expression}
\usage{
ggpacket_plus_ANY(e1, e2)
}
\description{
Swallow calls when a ggpacket is added to any expression
}
