% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcall.R
\name{infer_ggcall_name}
\alias{infer_ggcall_name}
\title{Label ggcall with function name if it can be deduced}
\usage{
infer_ggcall_name(expr)
}
\arguments{
\item{expr}{An expression from which a call name should be inferred.}
}
\description{
Label ggcall with function name if it can be deduced
}
\keyword{internal}
