% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_df.r
\name{last_args}
\alias{last_args}
\title{Filter argument list for only unnamed items and the last instance of each
named item}
\usage{
last_args(args)
}
\arguments{
\item{args}{a character vector of values}
}
\value{
a filtered vector of values with only the last instance of non-empty
strings as well as any empty strings ('') remaining
}
\description{
Filter argument list for only unnamed items and the last instance of each
named item
}
