% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpacket.R
\name{update_data}
\alias{update_data}
\title{Reduce data parameters, iteratively applying functions or masking}
\usage{
update_data(d1, d2, ...)
}
\arguments{
\item{d1}{A plot data object to update}

\item{d2}{A second plot data object with which to update \code{d1}}

\item{...}{Additional objects to sequentially collapse.}
}
\description{
Reduce data parameters, iteratively applying functions or masking
}
